pyproject.toml
[build-system]
requires = ["setuptools>=40"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-secretary-skills"
version = "0.1.0"
description = "AI Secretary Google Workspace skill dependencies"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.0",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.0",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "pytz>=2024.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*_service*", "google_auth*", "shared*"]
